+++ src/components/icon/index.tsx
@@ -78,9 +78,12 @@ import { Unlock_line } from "./unlock_line";
 import { Game_2_line } from "./game_2_line";
 import { Star_fill } from "./star_fill";
 import { Pencil_line } from "./pencil_line";
+import { Plugin_2_line } from "./plugin_2_line";
 
 const Icon: React.FC<IconProps> = (props) => {
   const { icon } = props;
+
+  if (icon === "plugin_2_line") return <Plugin_2_line {...props} />;
 
   if (icon === "pencil_line") return <Pencil_line {...props} />;