+++ src/hooks/usePlugin/types.ts
@@ -1,0 +1,11 @@
+type GoogleSearch = {
+  enable: boolean;
+  engine_id: string;
+  api_key: string;
+};
+
+export type PluginStore = {
+  google_search: GoogleSearch;
+
+  updateGoogleSearch: (google_search: GoogleSearch) => void;
+};