+++ src/lib/index.ts
@@ -4,3 +4,5 @@ export * from "./platform";
 export * from "./response";
 export * from "./calcTokens";
 export * from "./isMobile";
+export * from "./sleep";
+export * from "./function_call";